Initial implementation to find palindromes
It only tries to find the palindrome words as an initial step.

The search starts from the full string removing a character from
the end until getting a match, but keeping the full string for future
searches where a character from the front is removed until the minimum 3
characters needed for a palindrome is reached.

Once a match is found, that part of the string is ignored and a new
search starts with the remaining characters.

The search stops when the remaining string length is less than the
minimum palindrome length.

Use an implicit class so the function works as an String extension.