Find palindromes including index and length
Replace current string with begin and end indexes.

Using indexes mean a copy of the remaining string is not necessary and
the full value can be accessed directly getting the relevant substring
using the indexes.

The challenge to distinguish between a new search (removing a character)
and the end of the algorithm is sorted by changing the order of the
conditions. Firstly, check whether the remaining string length is less
than the minimum. If it is, finish the algorithm. Then, check the length
has reached the exact minimum only after we've check if the current
string is palindrome. If the minimum hasn't been reached yet, carry on
the current search.